Adding JUnit test suite
Added unit tests for all classes and their methods
Unfortunately, QuakeDataPrinterTest is not currently included in the
AllTests suite. I could not understand the problem with it. I tried
using several Text Diff applications, and all confirmed that the 2
strings were exactly the same - and yet assertEquals would not accept
them as the same text.

I may go back to this later, but for now I will focus on the Extra
Challenge, registering the Means of Death for the games.